Dummy commit.
This will have to be rebased so as to clean the mess.

In this commit are the settings which I have been using to try a
clean shutdown with Ctrl + C. The idea behind it is, launch
`python test.py` from a console and then kill it with Ctrl + C.

We can see that the daemon is the first one who gets the signal,
and right now it prevents the processes from getting it. By
reverting the changes in nameserver.py:92, we can see that the
problem is that we are not able to locate the NS after Ctrl + C,
so the NameServerProcess cannot know when all agents have been
successfully shutdown.

Further investigation is required.